hallucination-buster-rag/requirements.txt
faiss-cpu
numpy
sentence-transformers
streamlit
requests
tqdm
scikit-learn
rapidfuzz

